Adds coordinator.users module
